✨ <feat> Clean logout and display logout message

+++ src/main/java/projetb2/overlook_hotel/repository/HotelUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import projetb2.overlook_hotel.model.HotelUser;
 
 @Repository
-public interface HotelUserRepository extends JpaRepository<HotelUser, Long> {
+public interface HotelUserRepository extends JpaRepository<HotelUser, Integer> {
     /**
      * Finds a HotelUser by their email.
      *
